Add option NANOPB_PB2_TEMP_DIR to store nanopb_pb2.py in a temporary directory (#601)
When environment variable NANOPB_PB2_TEMP_DIR is defined, the generator will
store the autogenerated nanopb_pb2.py in a temporary directory. This can
be the value of the environment variable, if it is valid directory, or
otherwise system-wide temp directory is used.

This avoids polluting the source code folder with autogenerated file, and may avoid some version
incompatibility issues. By default the old behavior of storing nanopb_pb2.py under proto folder is
preserved.