Ensure generator files are updated when cached
When bumping Nanopb, while having the previous version cached, the
generator files are not correctly copied during build.

CMake copy_directory does not seem to copy the content if the directory
does already exist.

By specifying that we want to copy the content of the directory
we can avoid this issue.

Review: Add comment